services/model_service/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.32.0

# Database
psycopg2-binary==2.9.11

# ML
mlflow==2.21.3
scikit-learn==1.6.1
xgboost==2.1.4
pandas==2.2.3
numpy==2.2.3

# Validation
pydantic==2.10.0
